Update js/threesixty.js
Converted to plugin since I will likely use this JQuery plugin for ZBrush and Maya turntables as it is awesome to show them off on the web!  Added options to enable passing in url format string to plugin, along with frame start, total frames, frame padding, etc.  All options which are typical when exporting an image sequence from a 3D application